Add docker setup for arm32v6
It uses a docker image with alpine/node for arm32v6. Docker image has to
install git to avoid problems with husky (shouldn't be necessary in the
future). Image gets tagged for ghcr.io since docker registry is moving
there.

docker-compose file will probably be replaced with x86/amd64 in the
future. It does run under those systems but emulation is needed.