Add new color pallete for layout
This adds the pallete being discussed in #20. Some colors might not be
needed and we can just use foundation SASS helper functions to get other
shades of those colors.

Some other colors are no longer used and they need to be discarded
though they do not add up into CSS output if they are not used.