Adds state rendering for reduxjs
Right now there's really only one route and only one of our components
is hooked to the store state. getInitialState and the dataLoaders need
further work.

Webpack config was slightly altered here, webpack-dev-server swallows
everything into a virtual FS (kind of nice for SSDs) but nodejs needs
access to the server bundle so that needs to be written to disk. It also
adds the react-hot-loader webpack plugin (which is already part of RHL)
as it's needed for hooks support and some other interesting scenarios.