Moves favicon and changes HTML needle
Favicon is best served from root as favicon.ico since that's what most
browsers will do. This way we don't get a lot of "fake" 404 errors.

HTML production bundle is minified, so using the HTML comment as a
needle to replace it with the real script is not possible as comments
are stripped away.

The favicon was also missing from the production HTML template, this
commits adds it.